Adding support for "alternate" plugins
This adds the ability for plugins to be configured to transform
the _entire_ build, not just individual files.

Basically, if the plugin `Function` has an `alternate` property
that is truthy, it will be added to `Duo#altPlugins` instead of
`Duo#plugins`.

In `Duo#run()`, the altPlugins Ware instance will be applied to
the `results` object, where each one can modify the `code` or
`map` as needed before the final return.