consider a modified manifest the same as a modified file
Currently, if you change the version in your manifest, the newer
version will not be resolved/fetched unless you touch your files.

The change here now causes all files to be re-parsed for dependencies
when the manifest has changed. (ie: mtime differs)

This opens up the possibility for using the manifest for managing
versions for a project, while removing the need to delete `components`
in order to get updates!