fixing bug in local path resolve
if you have a local dep that points to a missing file, it will attempt to resolve it
as a remote. duo-parse treats something like `../../a/b/c` as `{ repo: '..' }`. if you
also have a manifest, it will attempt to find `..` via regexp in the `dependencies`
list. since it's a regexp, `..` will match just about anything.

the end result is that you have an extra dependency being added to your build. in the
case of css, it can be hard to trace this down.

this throws when the dep looks like a path (ie: begins with "." or "/") and is not
resolved locally. I don't believe it is possible to have a github repo that begins
with either of these chars, so we should be safe here.